Rename app build to buildhistory and reset migrations
